refactor: reworked use of unsigned integer types
This project exclusively uses fixed integer width types to align with
their prominent use in Vulkan APIs even in cases where its unnecessary.
This contradicts the Google C++ Style Guide which recommends using int
unless there is a specific reason to do so.

After reviewing the code, a decision was made to use int where possible
unless it adds undue burden when interacting with Vulkan APIs. This
change also includes minor changes to remove multiple signed/unsigned
integer comparisons.

Guidance on integer types in the Google C++ Style Guide can be found
here: https://google.github.io/styleguide/cppguide.html#Integer_Types

Closes #36

+++ src/engine/swapchain.cpp
@@ -55,8 +55,8 @@ vk::Extent2D GetSwapchainImageExtent(const gfx::Window& window,
   const auto [min_width, min_height] = surface_capabilities.minImageExtent;
   const auto [max_width, max_height] = surface_capabilities.maxImageExtent;
   const auto [framebuffer_width, framebuffer_height] = window.GetFramebufferSize();
-  return vk::Extent2D{.width = std::clamp(framebuffer_width, min_width, max_width),
-                      .height = std::clamp(framebuffer_height, min_height, max_height)};
+  return vk::Extent2D{.width = std::clamp(static_cast<std::uint32_t>(framebuffer_width), min_width, max_width),
+                      .height = std::clamp(static_cast<std::uint32_t>(framebuffer_height), min_height, max_height)};
 }
 
 std::vector<vk::UniqueImageView> CreateSwapchainImageViews(const vk::Device& device,
